+++ vpin-studio-server/src/test/java/de/mephisto/vpin/server/vpinmame/VPinMameServiceTest.java
@@ -15,7 +15,6 @@ import java.util.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)